Remove BookFormFields shared component
Each form now renders its own fields directly.
BookFormValues and bookFormSchema moved to bookFormSchema.ts.
